backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
aiohttp==3.9.1
numpy==1.24.3
scikit-learn==1.3.2
xgboost==2.0.3
python-dotenv==1.0.0

